requirements.txt
streamlit
openai
pandas
fpdf
pytesseract
pillow
